# backend/docs/swagger.yaml
definitions:
...
  github_com_diegoclair_leaderpro_internal_transport_rest_viewmodel.Login:
    properties:
      cpf:
        maxLength: 11
        minLength: 11
        type: string
      password:
        minLength: 8
        type: string
    required:
    - cpf
    - password
    type: object
  github_com_diegoclair_leaderpro_internal_transport_rest_viewmodel.LoginResponse:
    properties:
      access_token:
        type: string
      access_token_expires_at:
        type: string
      refresh_token:
        type: string
      refresh_token_expires_at:
        type: string
    type: object
...
  github_com_diegoclair_leaderpro_internal_transport_rest_viewmodel.RefreshTokenRequest:
    properties:
      refresh_token:
        type: string
    required:
    - refresh_token
    type: object
  github_com_diegoclair_leaderpro_internal_transport_rest_viewmodel.RefreshTokenResponse:
    properties:
      access_token:
        type: string
      access_token_expires_at:
        type: string
    type: object
...
  resterrors.restErr:
    properties:
      causes: {}
      error:
        type: string
      message:
        type: string
      status_code:
        type: integer
    type: object
host: localhost:5000
info:
  contact:
    email: diego93rodrigues@gmail.com
    name: Email
  description: This is an API documentation for go boilerplate
  title: Go Boilerplate API
  version: "1.0"
paths:
...
  /auth/login:
    post:
      consumes:
      - application/json
      description: Login
      parameters:
      - description: Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_diegoclair_leaderpro_internal_transport_rest_viewmodel.Login'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_diegoclair_leaderpro_internal_transport_rest_viewmodel.LoginResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/resterrors.restErr'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/resterrors.restErr'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/resterrors.restErr'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/resterrors.restErr'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/resterrors.restErr'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/resterrors.restErr'
      summary: Login
      tags:
      - auth
  /auth/logout:
    post:
      consumes:
      - application/json
      description: Logout the user
      parameters:
      - description: User access token
        in: header
        name: user-token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/resterrors.restErr'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/resterrors.restErr'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/resterrors.restErr'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/resterrors.restErr'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/resterrors.restErr'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/resterrors.restErr'
      summary: Logout
      tags:
      - auth
  /auth/refresh-token:
    post:
      consumes:
      - application/json
      description: Generate a new token using the refresh token
      parameters:
      - description: Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_diegoclair_leaderpro_internal_transport_rest_viewmodel.RefreshTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_diegoclair_leaderpro_internal_transport_rest_viewmodel.RefreshTokenResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/resterrors.restErr'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/resterrors.restErr'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/resterrors.restErr'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/resterrors.restErr'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/resterrors.restErr'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/resterrors.restErr'
      summary: Refresh Token
      tags:
      - auth
...
schemes:
- http
swagger: "2.0"